JP-3990: User is not notified that the communication schedule that is being deleted is still in use

+++ comu/com.energyict.mdc.scheduling.rest/src/main/java/com/energyict/mdc/scheduling/rest/impl/SchedulingResource.java
@@ -103,7 +103,12 @@ public class SchedulingResource {
     @Produces(MediaType.APPLICATION_JSON)
     public Response deleteSchedules(@PathParam("id") long id) {
         ComSchedule comSchedule = findComScheduleOrThrowException(id);
-        comSchedule.delete();
+        if (this.isInUse(comSchedule)) {
+            comSchedule.makeObsolete();
+        }
+        else {
+            comSchedule.delete();
+        }
         return Response.noContent().build();
     }
